refactor: extract to dedicated utils file
